Add xbox controller support and an error log folder with a bug I found
Now the game looks to see if there is an xbox controller hooked up and
uses vibration to alert the player of being hit. Might need to tweak the
vibration setting depending on how you guys feel about the time and
intensity of it.

Also, added an error log file that contains one bug that was found while
testing the game. I will next add the high score keeper file (maybe
Geoff can add a high score screen after player dies?) and a gauge that
works off the number of enemies shot.