Added capability to use a single powerup and did some refactoring
Added the ability to plug in a single powerup to use for the player
(time slowdown) and is triggered by the right trigger on a controller.

Also did some refactoring of the BaseEnemy class as it is should be an
abstract class and only has other Enemies inherit from it, also cleaned
up some of the accessibility of certain functions and variables.

Added exception message for the index out of bounds error to diagnose
what is happening/why we are getting the error.